Insure proper insert order by creating an atRule queue.
