+++ packages/emotion/test/css.test.js
@@ -101,46 +101,6 @@ describe('css', () => {
     expect(tree).toMatchSnapshot()
   })
 
-  test('handles media query merges', () => {
-    const mq = [
-      '@media(min-width: 420px)',
-      '@media(min-width: 640px)',
-      '@media(min-width: 960px)'
-    ]
-    const buttonCSS = [
-      {
-        color: 'red',
-        [mq[1]]: {
-          color: 'blue'
-        }
-      },
-      {
-        color: 'purple',
-        [mq[1]]: {
-          color: 'aquamarine'
-        }
-      }
-    ]
-
-    const cls1 = css([
-      {
-        color: 'darkslateblue',
-        [mq[0]]: {
-          color: 'amethyst'
-        },
-        [mq[1]]: {
-          color: 'rebeccapurple'
-        },
-        [mq[2]]: {
-          color: 'burlywood'
-        }
-      },
-      buttonCSS
-    ])
-    const tree = renderer.create(<div className={cls1} />).toJSON()
-    expect(tree).toMatchSnapshot()
-  })
-
   test('computed key is only dynamic', () => {
     const cls1 = css({
       fontSize: 10,
